ci(init,bootstrap): Improve resiliency of snapcraft.yamls
Building locally is brittle due to artefacts polluting the build
container (caused by `source: .`). This doesn't present in ci because
the build container is fresh. I did not see a way to improve
`snapcraft.yaml` to better handle host artefacts. So, to avoid confusion
this updates `ubuntu-bootstrap` and `ubuntu-desktop-init` parts to use
`source: <repo-url>` and `source-branch: ci/(init|bootstrap)`.

For local iterations, and if familiar with the gotcha's, these values
may be overridden or set to point to your own repo. Finally, cleaning a
local setup leads to slower builds but can be done with:

- `flutter clean` # .pub-cache can cause build to fail due to host refs
- `snapcraft clean` # tainted parts can cause subsequent runs to fail
- `rm -rf packages/subiquity_client/subuqity/.subiquity`
- `rm -rf snap`
- `./ci/snap-build-setup.sh (bootstrap|init)`